rename targets module to target
