+++ .gitignore
@@ -14,3 +14,5 @@ target/
 .DS_Store
 *.swp
 .idea/
+
+!espflash/src/target
