+++ espflash/src/lib.rs
@@ -36,7 +36,7 @@ pub use self::error::Error;
 pub mod connection;
 pub mod flasher;
 pub mod image_format;
-pub mod targets;
+pub mod target;
 
 mod error;
